refactor: draw.cの機能を切り分け。
#80
