+++ includes/thread.h
@@ -1,0 +1,11 @@
+#ifndef THREAD_H
+# define THREAD_H
+
+# include "data.h"
+# include "shading.h"
+# include "calc.h"
+# include <pthread.h>
+
+void	thread_store_canvas(t_data *data, char *canvas);
+
+#endif // THREAD_H
